refactor: util func to calculate center
